Support complex types via overlays
This commit is a proof of concept update of vss-tools to support
higher order/complex types via the approach outlined in
https://github.com/COVESA/vehicle_signal_specification/wiki/Proposal-%22struct%22:-%22Weak%22-layer-reference-to-external-IDLs

**Description  **
The main idea is that we are trying to model a sensor of Complex type - `ObstacleList` as defined in comments of
https://github.com/COVESA/vehicle_signal_specification/issues/326

For reference, the type is defined as follows:
```
ObstacleClass
{
    uint64 class
}

Obstacle
{
    uint64 id
    ObstacleClass class
    float64 probability
}

ObstacleList
{
    Obstacle[] obstacles
}
```

In the main vspec file (struct_overlay.json), the sensor's type is erased - we treat the sensor's data type as a byte-array.

The exact type information is then provided in an overlay file (obstacles_overlay.vspec) where the required information needed by custom tooling is provided .

In this example, it is assumed that custom tooling is available to process the following information:
1. datatypename - The actual data type name of the sensor
2. idl - IDL file defining the data type shape

The vss2json tool is updated to include these custom keys in the output. (out.json)

All the relevant files are included in the commit.

vspec -> json: `python vspec2x.py --format json Struct_overlay.vspec -o obstacles_overlay.vspec --json-pretty out.json`