Support complex types via branches
This commit is a proof of concept update of vss-tools to support
higher order/complex types via a variation of the approach outlined in https://github.com/COVESA/vehicle_signal_specification/wiki/Proposal-%22struct%22:-Just-model-as-branches,-the-rest-is-deployment

Ref: https://github.com/COVESA/vehicle_signal_specification/issues/326

**Description  **
The main idea is that we are trying to model a sensor of Complex type - `ObstacleList` as a branch.
The current limitation with branches is that you cannot have arrays of branches.

With the understanding that branches are to be treated as a data type (like primitive and their array equivalent),
what we need is a way to convey the fact that the branch represents a collection/array.
In the case of primitives, we use the operator `[]` for this purpose. For example, an array type `uint8` is denoted by adding the suffix `[]`.

On a similar path, we can utilize the already available `arraysize` property of a VSS node to specify the array size of a branch typed node.

In this commit, the VSS tools are updated to proceess the `arraysize` key for branch typed nodes.

For reference, the type is defined as follows:
```
ObstacleClass
{
    uint64 class
}

Obstacle
{
    uint64 id
    ObstacleClass class
    float64 probability
}

ObstacleList
{
    Obstacle[] data
}
```

All the relevant files are included in the commit.

vspec -> json: `python vspec2json.py ObstacleArray.vspec --json-pretty out.json`

The resulting JSON output contains the attribute "arraysize" for the branch typed nodes if defined.